# Optimize ToTitleCase Method for performance & allocation