@damieng 2024-10-05
We appreciate the discussion of performance and intent behind the PR but ToTitleCase is not actually used. It's a left-over helper method when we were considering providing a set of case conversions.

In the end only the camel case methods are used and not by default.  A user would have to manually register the CamelCaseElementNameConvention. Additionally these methods are only called once when the first DbContext is created and then the resultings strings are cached in the model metadata so any improvements to them would have negligble real-world implications.